+++ src/pages/Gen4DailyRNG/DailyRNGList/MtCoronetB1FMap/MapData.ts
