+++ src/pages/Gen4DailyRNG/DailyRNGList/define.ts
@@ -1,0 +1,7 @@
+export type DailyResult = {
+  seed: string
+  lottery: number
+  mapName: string
+  pokemon: string
+  points: readonly number[]
+}
