+++ src/pages/Gen4DailyRNG/DailyRNGList/index.ts
@@ -1,0 +1,1 @@
+export { DailyRNGListPage } from './Page'
