+++ src/pages/Gen4DailyRNG/DailyRNGPage.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Flex, Box, Button, HStack } from '@chakra-ui/react'
 
 import { DailySeedFinder } from './DailySeedFinder'
-import { DailyRNGList } from './DailyRNGList'
+import { DailyRNGListPage } from './DailyRNGList'
 
 export const DailyRNGPage: React.FC = () => {
   const [tab, setTab] = React.useState<'DailySeedFinder' | 'DailySeedList'>('DailySeedFinder')
@@ -36,7 +36,7 @@ export const DailyRNGPage: React.FC = () => {
         </HStack>
       </Flex>
       {tab === 'DailySeedFinder' && <DailySeedFinder />}
-      {tab === 'DailySeedList' && <DailyRNGList />}
+      {tab === 'DailySeedList' && <DailyRNGListPage />}
     </Box>
   )
 }
